rename icons-v2 folder to icons
