+++ components/icons/Exit.tsx
