+++ components/layout/Footer.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import { CSSProperties } from 'react';
 import styles from './Footer.module.css';
-import ChevronUp from '../icons-v2/ChevronUp';
+import ChevronUp from '../icons/ChevronUp';
 
 interface Props {
   style?: CSSProperties,
